pyproject.toml
[project]
name = "meeting-coach-agent"
version = "0.1.0"
description = "AI-powered meeting analysis agent using Claude Agent SDK and AWS Transcribe"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "claude-agent-sdk>=0.1.0",
    "boto3>=1.34.0",
    "python-dotenv>=1.0.0",
    "streamlit>=1.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

